requirements.txt
SQLAlchemy
Flask
Flask-Migrate
Flask-SQLAlchemy
flask-cors
python-dotenv